Refactoring completed
Loads of changes. Bins are now properties of classes rather than matrices which are passed around, temporary edges are deleted rather than detached if split, improved logic for deleting vertices and edges, and changed the functions in other files to reflect these changes, as well as prior directionality changes.

Also added functions to restore the network to the original (as in, after the initial network building has been done) version, as well as ones to check two states of the network and report differences between the two.

Moved lat, lon -> bin indices calculations to a misc file.

Lots of other stuff. Been coding for most of the day.

Deleted tests, because they need to be rewritten from scratch.