read_csv_stream sanity checks
Changed second in read_csv_stream from a list of integers to a list of strings. It looks through the first row of the csv file and looks for each included string, which are used to determine what is the timestamp, what is latitude, etc.

Also applying this plus a bit of tightening up in main and clean_trips.py.